requirements.txt
flask==3.0.3
anthropic==0.40.0
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
